third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_after_target_removed_from_slot_mouse-expected.txt

This is a testharness.js-based test.
[FAIL] Pointer events from mouse received before/after slot removal at pointerdown
  assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-removed),pointerout@child,pointerleave@child,pointerover@parent,pointerup@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-removed),pointerout@child,pointerleave@child,pointerover@parent,pointerenter@parent,pointerup@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent"
[FAIL] Pointer events from mouse received before/after slotted-child removal at pointerdown
  assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-removed),pointerover@parent,pointerup@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-removed),pointerover@parent,pointerenter@parent,pointerup@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent"
[FAIL] Pointer events from mouse received before/after slot removal at pointerup
  assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerout@child,pointerleave@child,pointerover@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerout@child,pointerleave@child,pointerover@parent,pointerenter@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent"
[FAIL] Pointer events from mouse received before/after slotted-child removal at pointerup
  assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerover@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerenter@parent,pointerover@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent"
Harness: the test ran to completion.
